Migration of /services endpoint using fastapi
