+++ wazo_sysconfd/plugins/systemd_services/services.py
@@ -7,24 +7,12 @@ import subprocess
 from xivo import http_json_server
 from xivo.http_json_server import HttpReqError
 from xivo.http_json_server import CMD_RW
+from . import exceptions
 
 logger = logging.getLogger('wazo_sysconfd.modules.services')
 
 
-class InvalidActionException(ValueError):
-    def __init__(self, service_name, action):
-        super(InvalidActionException, self).__init__(self)
-        self.service_name = service_name
-        self.action = action
-
-
-class InvalidServiceException(ValueError):
-    def __init__(self, service_name):
-        super(InvalidServiceException, self).__init__(self)
-        self.service_name = service_name
-
-
-def services(args, options):
+def services(args):
     """
     POST /services
 
@@ -42,16 +30,16 @@ def _run_action_for_service(service, action):
     try:
         _validate_action(service, action)
         output = _run_action_for_service_validated(service, action)
-    except InvalidActionException as e:
+    except exceptions.InvalidActionException as e:
         logger.error("action %s not authorized on %s service", e.action, e.service_name)
-    except InvalidServiceException as e:
+    except exceptions.InvalidServiceException as e:
         logger.error("service %s is not valid", e.service_name)
     return output
 
 
 def _validate_action(service_name, action):
     if action not in ['stop', 'start', 'restart']:
-        raise InvalidActionException(service_name, action)
+        raise exceptions.InvalidActionException(service_name, action)
 
 
 def _run_action_for_service_validated(service, action):
@@ -71,7 +59,10 @@ def _run_action_for_service_validated(service, action):
         logger.exception("Error while executing action")
         raise HttpReqError(500, "can't manage services")
 
-    return output
+    if isinstance(output, bytes):
+        return output.decode("utf-8")
+    else:
+        return output 
 
 
 http_json_server.register(services, CMD_RW, name='services')
